[OSD-21095-enable-optIn-region-for-new-accounts] moved GetFeatureFlagValue func to pkg/utils
